Refactoring to Hygraph
