+++ .prettierrc
@@ -1,5 +1,5 @@
 {
-	"useTabs": false,
+	"useTabs": true,
 	"singleQuote": true,
 	"trailingComma": "none",
 	"printWidth": 100,
