+++ src/app.html
@@ -1,20 +1,17 @@
 <!DOCTYPE html>
 <html lang="en">
-  <head>
-    <meta charset="utf-8" />
-    <meta name="description" content="" />
-    <link rel="icon" type="image/png" href="https://styleguide.fdnd.nl/favicon.png" />
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <link rel="preconnect" href="https://fonts.googleapis.com" />
-    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
-    <link
-      rel="stylesheet"
-      href="https://fonts.googleapis.com/css2?family=Open+Sans:ital,wght@0,300;0,400;0,600;1,300;1,400&display=swap"
-    />
-    <link rel="stylesheet" href="https://styleguide.fdnd.nl/fdnd.v1.css" />
-    %sveltekit.head%
-  </head>
-  <body class="white-on-blue">
-    %sveltekit.body%
-  </body>
+	<head>
+		<meta charset="utf-8" />
+		<meta name="description" content="" />
+		<link rel="icon" type="image/png" href="https://styleguide.fdnd.nl/favicon.png">
+		<meta name="viewport" content="width=device-width, initial-scale=1" />
+		<link rel="preconnect" href="https://fonts.googleapis.com" />
+		<link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
+		<link rel="stylesheet" href="https://fonts.googleapis.com/css2?family=Open+Sans:ital,wght@0,300;0,400;0,600;1,300;1,400&display=swap" />
+		<link rel="stylesheet" href="https://styleguide.fdnd.nl/fdnd.v1.css" />
+		%sveltekit.head%
+	</head>
+	<body class="white-on-blue">
+		%sveltekit.body%
+	</body>
 </html>
