--- src/error.html
@@ -1,6 +1,0 @@
-<svelte:head>
-	<title>%sveltekit.status%</title>
-</svelte:head>
-
-<h1>%sveltekit.status%</h1>
-<p>%sveltekit.error.message%</p>
