+++ src/routes/+layout.svelte
@@ -1,28 +1,28 @@
 <script>
-  import Nav from '$lib/components/Nav.svelte';
-  import Footer from '$lib/components/Footer.svelte';
-  import GAnalytics from '$lib/components/GAnalytics.svelte';
-
-  export const gtmId = 'G-QWP67QPK22';
+	import Nav from '$lib/components/Nav.svelte'
+	import Footer from '$lib/components/Footer.svelte'
 </script>
 
 <svelte:head>
-  <title>FDND Programma</title>
+	<title>FDND Programma</title>
 </svelte:head>
 
-<GAnalytics {gtmId} />
-
 <header>
-  <h1><span>FDND</span> Programma</h1>
+	<h1><span>FDND</span> Programma</h1>
 </header>
 
 <Nav />
 
 <main>
-  <slot />
+	<slot />
 </main>
 
 <Footer />
 
 <style>
+	@media (min-width: 30rem) {
+		h1 {
+			margin-left: 2rem;
+		}
+	}
 </style>
