+++ src/routes/+page.server.js
@@ -1,29 +1,30 @@
-import createClient from '$lib/vendor/prismicio'
-import { error } from '@sveltejs/kit'
+import { client } from '$lib/utils/client'
 
-export async function load({ fetch, request }) {
-  const client = createClient({ fetch, request })
-  const page = await client.getSingle('program')
-  const semesters = await Promise.all(
-    page.data.semesters.map(async (s) => {
-      const semester = await client.getByUID('semester', s.semester.uid)
-      const sprints = await Promise.all(
-        semester.data.sprints.map(async (s) => {
-          const sprint = await client.getByUID('sprint', s.sprint.uid)
-          return sprint
-        })
-      )
-      semester.data.sprints = sprints
-      return semester
-    })
-  )
+export const load = async () => {
+	const query = `
+    query Programs {
+        program(where: {id: "cl9pm4is9l9xu0atdbqxnd710"}) {
+            title
+            subtitle
+            content {
+                html
+            }
+            semesters {
+                title
+                slug
+                sprints {
+                    title
+                    slug
+                    sprintNumber
+                    startdate
+                }
+            }
+        }
+    }`
 
-  if (page && semesters) {
+    const data = await client({ query, fetch: fetch });
+
     return {
-        page,
-        semesters
-      } 
-  }
-
-  throw error(404, 'Not found')
+        ...data.program
+    }
 }
