+++ vite.config.js
@@ -1,6 +1,5 @@
 import { sveltekit } from '@sveltejs/kit/vite';
 
-/** @type {import('vite').UserConfig} */
 const config = {
 	plugins: [sveltekit()]
 };
